# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Generate pay slip",
    "description": "version 2.0",
    "version": "1.0.0"
  },
  "host": "pay-slip-1.onrender.com",
  "basePath": "/",
  "schemes": [
    "https",
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/check/{token}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}